refactor(mmrelay.meshtastic): extract connection/events to modules
move connection-related helpers and event handlers from meshtastic_utils
into new modules (connection.py and events.py). update imports to expose
APIs from the new modules and remove redundant in-file implementations.
this decouples core utilities from meshtastic integrations and enables
targeted testing and maintenance
