+++ src/mmrelay/meshtastic_utils.py
@@ -139,6 +139,20 @@ from mmrelay.meshtastic.ble import (
     _scan_for_ble_address,
     _validate_ble_connection_address,
 )
+from mmrelay.meshtastic.connection import (
+    _connect_meshtastic_impl,
+    _get_connect_time_probe_settings,
+    _get_connection_retry_wait_time,
+    _rollback_connect_attempt_state,
+    _schedule_connect_time_calibration_probe,
+    connect_meshtastic,
+    serial_port_exists,
+)
+from mmrelay.meshtastic.events import (
+    on_lost_meshtastic_connection,
+    on_meshtastic_message,
+    reconnect,
+)
 from mmrelay.meshtastic.executors import (
     _clear_ble_future,
     _clear_metadata_future_if_current,
@@ -1160,2093 +1174,10 @@ def _get_device_metadata(
 # for backward-compatible patch targets (tests patch mmrelay.meshtastic_utils.*).
 
 
-def serial_port_exists(port_name: str) -> bool:
-    """
-    Determine whether a serial port with the given device name exists on the system.
-
-    Parameters:
-        port_name (str): Device name to check (e.g., '/dev/ttyUSB0' on Unix or 'COM3' on Windows).
-
-    Returns:
-        `True` if a matching port device name is present, `False` otherwise.
-    """
-    ports = [p.device for p in serial.tools.list_ports.comports()]
-    return port_name in ports
-
-
-def _get_connection_retry_wait_time(attempts: int) -> float:
-    """Return capped exponential retry backoff without exponentiating past the cap."""
-    if attempts <= 0 or CONNECTION_RETRY_BACKOFF_MAX_SECS <= 0:
-        return 0.0
-
-    if CONNECTION_RETRY_BACKOFF_BASE <= 1:
-        return min(
-            float(CONNECTION_RETRY_BACKOFF_BASE**attempts),
-            float(CONNECTION_RETRY_BACKOFF_MAX_SECS),
-        )
-
-    max_capped_attempt = math.ceil(
-        math.log(
-            CONNECTION_RETRY_BACKOFF_MAX_SECS,
-            CONNECTION_RETRY_BACKOFF_BASE,
-        )
-    )
-    exponent = min(attempts, max_capped_attempt)
-    return min(
-        float(CONNECTION_RETRY_BACKOFF_BASE**exponent),
-        float(CONNECTION_RETRY_BACKOFF_MAX_SECS),
-    )
-
-
-def _get_connect_time_probe_settings(
-    active_config: dict[str, Any] | None, connection_type: str
-) -> tuple[bool, float]:
-    """
-    Return connect-time probe enablement and timeout settings for a connection.
-    """
-    default_timeout = float(DEFAULT_MESHTASTIC_OPERATION_TIMEOUT)
-    default_enabled = DEFAULT_HEALTH_CHECK_ENABLED
-    if not isinstance(active_config, dict):
-        return default_enabled, default_timeout
-
-    meshtastic_cfg = active_config.get(CONFIG_SECTION_MESHTASTIC)
-    if not isinstance(meshtastic_cfg, dict):
-        return default_enabled, default_timeout
-
-    health_cfg = meshtastic_cfg.get(CONFIG_KEY_HEALTH_CHECK)
-    if not isinstance(health_cfg, dict):
-        return default_enabled, default_timeout
-
-    inherited_enabled = _coerce_bool(
-        health_cfg.get(CONFIG_KEY_ENABLED, default_enabled),
-        default_enabled,
-        "meshtastic.health_check.enabled",
-    )
-    enabled = _coerce_bool(
-        health_cfg.get(CONFIG_KEY_CONNECT_PROBE_ENABLED, inherited_enabled),
-        inherited_enabled,
-        "meshtastic.health_check.connect_probe_enabled",
-    )
-    timeout_secs = _coerce_positive_float(
-        health_cfg.get(CONFIG_KEY_PROBE_TIMEOUT, default_timeout),
-        default_timeout,
-        "meshtastic.health_check.probe_timeout",
-    )
-    return enabled, timeout_secs
-
-
-def _schedule_connect_time_calibration_probe(
-    client: Any,
-    *,
-    connection_type: str,
-    active_config: dict[str, Any] | None,
-) -> None:
-    """
-    Best-effort one-shot metadata probe after connect for skew calibration backup.
-    """
-    enabled, timeout_secs = _get_connect_time_probe_settings(
-        active_config, connection_type
-    )
-    if not enabled:
-        logger.debug("Connect-time metadata probe is disabled in configuration")
-        return
-
-    local_node = getattr(client, "localNode", None)
-    if local_node is None or not callable(getattr(client, "sendData", None)):
-        logger.debug(
-            "Skipping connect-time metadata probe; client lacks localNode/sendData support"
-        )
-        return
-
-    try:
-        probe_future = _submit_metadata_probe(
-            functools.partial(
-                _probe_device_connection,
-                client,
-                timeout_secs,
-            )
-        )
-    except MetadataExecutorDegradedError:
-        logger.debug(
-            "Skipping connect-time metadata probe; metadata executor is degraded"
-        )
-        return
-    except RuntimeError as exc:
-        logger.debug(
-            "Skipping connect-time metadata probe; submission failed",
-            exc_info=exc,
-        )
-        return
-
-    if probe_future is None:
-        logger.debug(
-            "Skipping connect-time metadata probe; metadata probe already in progress"
-        )
-        return
-
-    logger.debug(
-        "Scheduled one-shot connect-time metadata probe (timeout=%.1fs)",
-        timeout_secs,
-    )
-
-
-def _rollback_connect_attempt_state(
-    client: Any,
-    client_assigned_for_this_connect: bool,
-    startup_drain_armed_for_this_connect: bool,
-    startup_drain_applied_for_this_connect: bool,
-    reconnect_bootstrap_armed_for_this_connect: bool,
-    lock_held: bool = False,
-) -> bool:
-    """
-    Centralize cleanup of partially-assigned clients and timing state when a connect attempt fails.
-
-    Returns the updated value for client_assigned_for_this_connect (always False after cleanup).
-    """
-    global meshtastic_client, meshtastic_iface, _relay_active_client_id
-    global _relay_startup_drain_deadline_monotonic_secs, _startup_packet_drain_applied
-    global _relay_reconnect_prestart_bootstrap_deadline_monotonic_secs
-
-    _lock_ctx = contextlib.nullcontext() if lock_held else meshtastic_lock
-    if client is not None and (
-        client_assigned_for_this_connect or client is meshtastic_iface
-    ):
-        with _lock_ctx:
-            if meshtastic_client is client or client is meshtastic_iface:
-                try:
-                    if client is meshtastic_iface:
-                        _disconnect_ble_interface(
-                            meshtastic_iface, reason="connect setup failed"
-                        )
-                        meshtastic_iface = None
-                    else:
-                        client.close()
-                except Exception as cleanup_error:  # noqa: BLE001 - best-effort cleanup
-                    logger.warning(
-                        "Error closing Meshtastic client after setup failure: %s",
-                        cleanup_error,
-                    )
-                finally:
-                    if meshtastic_client is client:
-                        meshtastic_client = None
-                        _relay_active_client_id = None
-
-    if (
-        startup_drain_armed_for_this_connect
-        or reconnect_bootstrap_armed_for_this_connect
-    ):
-        with _relay_rx_time_clock_skew_lock:
-            if startup_drain_armed_for_this_connect:
-                _relay_startup_drain_deadline_monotonic_secs = None
-                if startup_drain_applied_for_this_connect:
-                    _startup_packet_drain_applied = False
-            if reconnect_bootstrap_armed_for_this_connect:
-                _relay_reconnect_prestart_bootstrap_deadline_monotonic_secs = None
-
-    return False
-
-
-def connect_meshtastic(
-    passed_config: dict[str, Any] | None = None,
-    force_connect: bool = False,
-) -> Any:
-    """
-    Establish a Meshtastic connection while preventing overlapping attempts.
-
-    This wrapper coordinates concurrent callers with an in-progress marker.
-    Callers that arrive while another connect is running wait for that attempt
-    to finish, then retry acquisition.
-    """
-    global _connect_attempt_in_progress
-
-    wait_budget_secs = _CONNECT_ATTEMPT_WAIT_MAX_SECS
-    config_source = (
-        passed_config
-        if isinstance(passed_config, dict)
-        else config if isinstance(config, dict) else None
-    )
-    if isinstance(config_source, dict):
-        meshtastic_section = config_source.get(CONFIG_SECTION_MESHTASTIC)
-        if isinstance(meshtastic_section, dict):
-            connection_type = meshtastic_section.get(CONFIG_KEY_CONNECTION_TYPE)
-            if connection_type == CONNECTION_TYPE_BLE:
-                wait_budget_secs = max(
-                    wait_budget_secs,
-                    _CONNECT_ATTEMPT_BLE_WAIT_MAX_SECS,
-                )
-
-    wait_deadline = time.monotonic() + wait_budget_secs
-
-    while True:
-        with _connect_attempt_condition:
-            if not _connect_attempt_in_progress:
-                _connect_attempt_in_progress = True
-                break
-
-            remaining_wait = wait_deadline - time.monotonic()
-            if remaining_wait <= 0:
-                logger.debug(
-                    "Timed out waiting for active connect attempt; returning no client"
-                )
-                return None
-
-            logger.debug(
-                "connect_meshtastic() already in progress; waiting for active attempt to finish"
-            )
-            while _connect_attempt_in_progress and not shutting_down:
-                remaining_wait = wait_deadline - time.monotonic()
-                if remaining_wait <= 0:
-                    break
-                _connect_attempt_condition.wait(
-                    timeout=min(_CONNECT_ATTEMPT_WAIT_POLL_SECS, remaining_wait)
-                )
-            if shutting_down:
-                logger.debug("Shutdown in progress. Not attempting to connect.")
-                return None
-            if _connect_attempt_in_progress and time.monotonic() >= wait_deadline:
-                logger.debug(
-                    "Timed out waiting for active connect attempt; returning no client"
-                )
-                return None
-
-    try:
-        return _connect_meshtastic_impl(
-            passed_config=passed_config,
-            force_connect=force_connect,
-        )
-    finally:
-        with _connect_attempt_condition:
-            _connect_attempt_in_progress = False
-            _connect_attempt_condition.notify_all()
-
-
-def _connect_meshtastic_impl(
-    passed_config: dict[str, Any] | None = None,
-    force_connect: bool = False,
-) -> Any:
-    """
-    Establishes a Meshtastic client connection using the configured connection type (serial, BLE, or TCP).
-
-    On success updates the module-level client state (meshtastic_client), may update matrix_rooms when a config is provided, and subscribes to meshtastic receive and connection-lost events once for the process lifetime. Honors shutdown and reconnect state and will respect `force_connect` to replace an existing connection.
-
-    Parameters:
-        passed_config (dict[str, Any] | None): Optional configuration to use in place of the module-level config; if provided and contains "matrix_rooms", that value will be used to update module-level matrix_rooms.
-        force_connect (bool): If True, forces creating a new connection even if a client already exists.
-
-    Returns:
-        The connected Meshtastic client instance on success, or `None` if a connection could not be established or shutdown is in progress.
-    """
-    global meshtastic_client, meshtastic_iface, shutting_down, reconnecting, config
-    global _relay_active_client_id
-    global RELAY_START_TIME, _relay_connection_started_monotonic_secs
-    global _relay_rx_time_clock_skew_secs, _relay_startup_drain_deadline_monotonic_secs
-    global _relay_reconnect_prestart_bootstrap_deadline_monotonic_secs
-    global _startup_packet_drain_applied
-    global matrix_rooms, _ble_future, _ble_future_address
-    global _ble_future_started_at, _ble_future_timeout_secs
-    if shutting_down:
-        logger.debug("Shutdown in progress. Not attempting to connect.")
-        return None
-
-    if reconnecting and not force_connect:
-        logger.debug("Reconnection already in progress. Not attempting new connection.")
-        return None
-
-    # Update the global config if a config is passed
-    if passed_config is not None:
-        config = passed_config
-
-        # If config is valid, extract matrix_rooms
-        if config and "matrix_rooms" in config:
-            matrix_rooms = config["matrix_rooms"]
-
-    with meshtastic_lock:
-        if meshtastic_client and not force_connect:
-            return meshtastic_client
-
-        # Close previous connection if exists
-        if meshtastic_client:
-            try:
-                if meshtastic_client is meshtastic_iface:
-                    # BLE needs an explicit disconnect to release BlueZ state; a
-                    # plain close() can leave the adapter "busy" for the next
-                    # connect attempt.
-                    _disconnect_ble_interface(meshtastic_iface, reason="reconnect")
-                    meshtastic_iface = None
-                else:
-                    meshtastic_client.close()
-            except Exception as e:
-                logger.warning(
-                    "Error closing previous connection: %s", e, exc_info=True
-                )
-            meshtastic_client = None
-            _relay_active_client_id = None
-
-        # Check if config is available
-        if config is None:
-            logger.error("No configuration available. Cannot connect to Meshtastic.")
-            return None
-
-        # Check if meshtastic config section exists
-        if (
-            CONFIG_SECTION_MESHTASTIC not in config
-            or config[CONFIG_SECTION_MESHTASTIC] is None
-        ):
-            logger.error(
-                "No Meshtastic configuration section found. Cannot connect to Meshtastic."
-            )
-            return None
-
-        # Check if connection_type is specified
-        if (
-            CONFIG_KEY_CONNECTION_TYPE not in config[CONFIG_SECTION_MESHTASTIC]
-            or config[CONFIG_SECTION_MESHTASTIC][CONFIG_KEY_CONNECTION_TYPE] is None
-        ):
-            logger.error(
-                "No connection type specified in Meshtastic configuration. Cannot connect to Meshtastic."
-            )
-            return None
-
-        # Determine connection type and attempt connection
-        connection_type = config[CONFIG_SECTION_MESHTASTIC][CONFIG_KEY_CONNECTION_TYPE]
-
-        # Support legacy "network" connection type (now "tcp")
-        if connection_type == CONNECTION_TYPE_NETWORK:
-            connection_type = CONNECTION_TYPE_TCP
-            logger.warning(
-                "Using 'network' connection type (legacy). 'tcp' is now the preferred name and 'network' will be deprecated in a future version."
-            )
-
-    # Move retry loop outside the lock to prevent blocking other threads
-    meshtastic_settings = config.get(CONFIG_SECTION_MESHTASTIC, {}) if config else {}
-    retry_limit_raw = meshtastic_settings.get("retries")
-    if retry_limit_raw is None:
-        retry_limit_raw = meshtastic_settings.get("retry_limit", INFINITE_RETRIES)
-        if "retry_limit" in meshtastic_settings:
-            logger.warning(
-                "'retry_limit' is deprecated in meshtastic config; use 'retries' instead"
-            )
-    try:
-        retry_limit = int(retry_limit_raw)
-    except (TypeError, ValueError):
-        retry_limit = INFINITE_RETRIES
-    attempts = 0
-    timeout_attempts = 0
-    successful = False
-    ble_scan_after_failure = False
-    ble_scan_reason: str | None = None
-
-    # Get timeout configuration (default: DEFAULT_MESHTASTIC_TIMEOUT)
-    timeout_raw = meshtastic_settings.get(
-        CONFIG_KEY_TIMEOUT, DEFAULT_MESHTASTIC_TIMEOUT
-    )
-    try:
-        timeout = int(timeout_raw)
-        if timeout <= 0:
-            logger.warning(
-                "Non-positive meshtastic.timeout value %r; using %ss fallback.",
-                timeout_raw,
-                DEFAULT_MESHTASTIC_TIMEOUT,
-            )
-            timeout = DEFAULT_MESHTASTIC_TIMEOUT
-    except (TypeError, ValueError):
-        # None or invalid value - use default silently
-        if timeout_raw is not None:
-            logger.warning(
-                "Invalid meshtastic.timeout value %r; using %ss fallback.",
-                timeout_raw,
-                DEFAULT_MESHTASTIC_TIMEOUT,
-            )
-        timeout = DEFAULT_MESHTASTIC_TIMEOUT
-    configured_timeout_secs = float(timeout)
-    configured_timeout_arg = max(1, math.ceil(configured_timeout_secs))
-    create_timeout_floor_secs = _coerce_positive_float(
-        _ble_interface_create_timeout_secs,
-        BLE_INTERFACE_CREATE_TIMEOUT_FLOOR_SECS,
-        "_ble_interface_create_timeout_secs",
-    )
-    ble_create_timeout_secs = max(configured_timeout_secs, create_timeout_floor_secs)
-
-    while (
-        not successful
-        and (retry_limit == 0 or attempts <= retry_limit)
-        and not shutting_down
-    ):
-        # Initialize before try block to avoid unbound variable errors
-        ble_address: str | None = None
-        supports_auto_reconnect = False
-        ble_connect_timeout_logged_for_attempt = False
-        startup_drain_pending_for_this_connect = False
-        startup_drain_armed_for_this_connect = False
-        startup_drain_applied_for_this_connect = False
-        reconnect_bootstrap_armed_for_this_connect = False
-        client_assigned_for_this_connect = False
-
-        client = None
-        try:
-            if connection_type == CONNECTION_TYPE_SERIAL:
-                # Serial connection
-                serial_port = config["meshtastic"].get(CONFIG_KEY_SERIAL_PORT)
-                if not serial_port:
-                    logger.error(
-                        "No serial port specified in Meshtastic configuration."
-                    )
-                    return None
-
-                logger.info(f"Connecting to serial port {serial_port}")
-
-                # Check if serial port exists before connecting
-                if not serial_port_exists(serial_port):
-                    raise serial.SerialException(
-                        f"Serial port {serial_port} does not exist."
-                    )
-
-                client = meshtastic.serial_interface.SerialInterface(
-                    serial_port, timeout=configured_timeout_arg
-                )
-
-            elif connection_type == CONNECTION_TYPE_BLE:
-                # BLE connection
-                ble_address = config["meshtastic"].get(CONFIG_KEY_BLE_ADDRESS)
-                if ble_address:
-                    logger.info(f"Connecting to BLE address {ble_address}")
-
-                    iface = None
-                    supports_auto_reconnect = False
-                    late_creation_disposer_future: Future[Any] | None = None
-                    with meshtastic_iface_lock:
-                        # If BLE address has changed, re-create the interface
-                        if (
-                            meshtastic_iface
-                            and getattr(meshtastic_iface, "address", None)
-                            != ble_address
-                        ):
-                            old_address = getattr(
-                                meshtastic_iface, "address", "unknown"
-                            )
-                            logger.info(
-                                f"BLE address has changed from {old_address} to {ble_address}. "
-                                "Disconnecting old interface and creating new one."
-                            )
-                            # Properly disconnect the old interface to ensure sequential connections
-                            _disconnect_ble_interface(
-                                meshtastic_iface, reason="address change"
-                            )
-                            meshtastic_iface = None
-
-                        if meshtastic_iface is None:
-                            # Disconnect any stale BlueZ connection before creating new interface
-                            _disconnect_ble_by_address(ble_address)
-
-                            # Create a single BLEInterface instance for process lifetime
-                            sanitized_address = _sanitize_ble_address(ble_address)
-                            logger.debug(
-                                f"Creating new BLE interface for {ble_address} (sanitized: {sanitized_address})"
-                            )
-                            # Detect whether this BLEInterface implementation supports
-                            # explicit auto_reconnect control.
-                            ble_init_sig = inspect.signature(
-                                meshtastic.ble_interface.BLEInterface.__init__
-                            )
-                            ble_kwargs = {
-                                "address": ble_address,
-                                "noProto": False,
-                                "debugOut": None,
-                                "noNodes": False,
-                                # Preserve user-configured Meshtastic reply timeout.
-                                "timeout": configured_timeout_arg,
-                            }
-
-                            # Configure auto_reconnect only when supported.
-                            supports_auto_reconnect = (
-                                "auto_reconnect" in ble_init_sig.parameters
-                            )
-                            if supports_auto_reconnect:
-                                ble_kwargs["auto_reconnect"] = False
-                                logger.debug(
-                                    "BLEInterface supports auto_reconnect; setting auto_reconnect=False "
-                                    "to ensure sequential reconnection control"
-                                )
-                            else:
-                                logger.debug(
-                                    "BLEInterface auto_reconnect parameter not available; using compatibility mode"
-                                )
-                            # Compatibility mode (official library) can benefit from
-                            # pre-scan retries. Forked interfaces that expose
-                            # auto_reconnect already perform richer connect/discovery
-                            # orchestration, so avoid duplicate scan work there.
-                            if ble_scan_after_failure and not supports_auto_reconnect:
-                                scan_timeout_secs = _coerce_positive_float(
-                                    _ble_scan_timeout_secs,
-                                    BLE_SCAN_TIMEOUT_SECS,
-                                    "_ble_scan_timeout_secs",
-                                )
-                                logger.debug(
-                                    "Scanning for BLE device before retrying %s (%s)",
-                                    ble_address,
-                                    ble_scan_reason or "previous failure",
-                                )
-                                _scan_for_ble_address(ble_address, scan_timeout_secs)
-                                ble_scan_after_failure = False
-                                ble_scan_reason = None
-
-                            # Create BLE interface with timeout protection to prevent indefinite hangs
-                            # Use ThreadPoolExecutor to run with timeout, as BLEInterface.__init__
-                            # can potentially block indefinitely if BlueZ is in a bad state.
-                            def create_ble_interface(
-                                kwargs: dict[str, Any],
-                            ) -> Any:
-                                """
-                                Create a BLEInterface configured for Meshtastic BLE connections.
-
-                                Parameters:
-                                    kwargs (dict): Keyword arguments forwarded to the Meshtastic BLEInterface constructor (e.g., `address`, `adapter`, `auto_reconnect`, `timeout`). Valid keys depend on the Meshtastic BLEInterface implementation.
-
-                                Returns:
-                                    BLEInterface: A newly constructed Meshtastic BLEInterface instance.
-                                """
-                                return meshtastic.ble_interface.BLEInterface(**kwargs)
-
-                            # Use the larger of configured connect timeout and the safety floor.
-                            # This keeps stale-worker detection and future.result() budget aligned
-                            # with the actual BLEInterface constructor timeout.
-                            create_timeout_secs = ble_create_timeout_secs
-
-                            # Guard against overlapping BLE tasks: if a previous BLE operation is
-                            # still running (often due to a hung BlueZ/DBus call), we skip queuing
-                            # a new task. Raising TimeoutError here intentionally reuses the
-                            # existing retry/backoff logic rather than silently proceeding.
-                            if shutting_down:
-                                logger.debug(
-                                    "Skipping BLE interface creation for %s (shutting down)",
-                                    ble_address,
-                                )
-                                raise TimeoutError(
-                                    f"BLE interface creation cancelled for {ble_address} (shutting down)."
-                                )
-
-                            _ensure_ble_worker_available(
-                                ble_address,
-                                operation="interface creation",
-                            )
-
-                            try:
-                                with _ble_executor_lock:
-                                    if _ble_future and not _ble_future.done():
-                                        logger.debug(
-                                            "BLE worker busy; skipping interface creation for %s",
-                                            ble_address,
-                                        )
-                                        raise TimeoutError(
-                                            f"BLE interface creation already in progress for {ble_address}."
-                                        )
-                                    if ble_address in _ble_executor_degraded_addresses:
-                                        logger.error(
-                                            "BLE executor degraded for %s: too many orphaned workers. "
-                                            "Reconnect or restart required to restore BLE connectivity.",
-                                            ble_address,
-                                        )
-                                        raise BleExecutorDegradedError(
-                                            f"BLE executor degraded for {ble_address}; reset required"
-                                        )
-                                    try:
-                                        future = _get_ble_executor().submit(
-                                            create_ble_interface, ble_kwargs
-                                        )
-                                    except RuntimeError as exc:
-                                        # The shared executor can be shutting down during interpreter
-                                        # teardown; treat this as a timeout so retry logic applies.
-                                        logger.exception(
-                                            "BLE interface creation submission failed for %s",
-                                            ble_address,
-                                        )
-                                        raise TimeoutError(
-                                            f"BLE interface creation could not be scheduled for {ble_address}."
-                                        ) from exc
-                                    _ble_future = future
-                                    _ble_future_address = ble_address
-                                    _ble_future_started_at = time.monotonic()
-                                    _ble_future_timeout_secs = create_timeout_secs
-                                future.add_done_callback(_clear_ble_future)
-                                try:
-                                    meshtastic_iface = (
-                                        _wait_for_future_result_with_shutdown(
-                                            future,
-                                            timeout_seconds=create_timeout_secs,
-                                        )
-                                    )
-                                    if meshtastic_iface is None:
-                                        _clear_ble_future(future)
-                                        raise RuntimeError(
-                                            "BLE interface creation returned no interface"
-                                        )
-                                    logger.debug(
-                                        f"BLE interface created successfully for {ble_address}"
-                                    )
-                                    if hasattr(meshtastic_iface, "auto_reconnect"):
-                                        supports_auto_reconnect = True
-                                    else:
-                                        reset_executor_degraded_state(
-                                            ble_address=ble_address
-                                        )
-                                except FuturesTimeoutError as err:
-                                    logger.error(
-                                        "BLE interface creation timed out after %.1f seconds for %s.",
-                                        create_timeout_secs,
-                                        ble_address,
-                                        exc_info=True,
-                                    )
-                                    logger.warning(
-                                        "This may indicate a stale BlueZ connection or Bluetooth adapter issue."
-                                    )
-                                    logger.warning(
-                                        BLE_TROUBLESHOOTING_GUIDANCE.format(
-                                            ble_address=ble_address
-                                        )
-                                    )
-                                    # Best-effort cancellation: if the worker is hung we cannot force
-                                    # it to stop, but this signals intent and lets retries proceed
-                                    # only if the future transitions to done/cancelled.
-                                    if future.cancel():
-                                        _clear_ble_future(future)
-                                    else:
-                                        _schedule_ble_future_cleanup(
-                                            future,
-                                            ble_address,
-                                            reason="interface creation timeout",
-                                        )
-                                        late_creation_disposer_future = future
-                                        timeout_count = _record_ble_timeout(ble_address)
-                                        _maybe_reset_ble_executor(
-                                            ble_address, timeout_count
-                                        )
-                                    meshtastic_iface = None
-                                    raise TimeoutError(
-                                        f"BLE connection attempt timed out for {ble_address}."
-                                    ) from err
-                            except TimeoutError as err:
-                                if shutting_down or str(err) == "Shutdown in progress":
-                                    if future.cancel():
-                                        _clear_ble_future(future)
-                                    else:
-                                        _schedule_ble_future_cleanup(
-                                            future,
-                                            ble_address,
-                                            reason="interface creation shutdown cancellation",
-                                        )
-                                        _attach_late_ble_interface_disposer(
-                                            future,
-                                            ble_address,
-                                            reason="interface creation shutdown cancellation",
-                                        )
-                                    meshtastic_iface = None
-                                raise
-                            except Exception:
-                                # Late BLE worker failures can surface during shutdown
-                                # after cancellation. Treat those as expected noise.
-                                if shutting_down:
-                                    logger.debug(
-                                        "BLE interface creation ended during shutdown for %s",
-                                        ble_address,
-                                        exc_info=True,
-                                    )
-                                else:
-                                    logger.exception("BLE interface creation failed")
-                                raise
-                        else:
-                            logger.debug(
-                                f"Reusing existing BLE interface for {ble_address}"
-                            )
-                            if hasattr(meshtastic_iface, "auto_reconnect"):
-                                supports_auto_reconnect = True
-                            else:
-                                try:
-                                    existing_sig = inspect.signature(
-                                        type(meshtastic_iface).__init__
-                                    )
-                                    supports_auto_reconnect = (
-                                        "auto_reconnect" in existing_sig.parameters
-                                    )
-                                except (TypeError, ValueError):
-                                    supports_auto_reconnect = False
-
-                        iface = meshtastic_iface
-
-                    if late_creation_disposer_future is not None:
-                        _attach_late_ble_interface_disposer(
-                            late_creation_disposer_future,
-                            ble_address,
-                            reason="interface creation timeout",
-                        )
-
-                    # Connect outside singleton-creation lock to avoid blocking other threads.
-                    # Interfaces that expose auto_reconnect support use explicit connect()
-                    # here; compatibility mode relies on constructor-managed connection.
-                    if (
-                        iface is not None
-                        and supports_auto_reconnect
-                        and hasattr(iface, "connect")
-                    ):
-                        logger.info(
-                            f"Initiating BLE connection to {ble_address} (sequential mode)"
-                        )
-
-                        # Add timeout protection for connect() call to prevent indefinite hangs
-                        # Use ThreadPoolExecutor with 30-second timeout (same as CONNECTION_TIMEOUT)
-                        def connect_iface(iface_param: Any) -> None:
-                            """
-                            Establishes the given interface by invoking its no-argument `connect()` method.
-
-                            Parameters:
-                                iface_param (Any): An interface-like object whose `connect()` method will be called to open the underlying connection.
-                            """
-                            iface_param.connect()
-
-                        # Check if shutting down before submitting connect() tasks
-                        if shutting_down:
-                            logger.debug(
-                                "Skipping BLE connect() for %s (shutting down)",
-                                ble_address,
-                            )
-                            raise TimeoutError(
-                                f"BLE connect cancelled for {ble_address} (shutting down)."
-                            )
-
-                        _ensure_ble_worker_available(
-                            ble_address,
-                            operation="connect",
-                        )
-
-                        with _ble_executor_lock:
-                            if _ble_future and not _ble_future.done():
-                                logger.debug(
-                                    "BLE worker busy; skipping connect() for %s",
-                                    ble_address,
-                                )
-                                raise TimeoutError(
-                                    f"BLE connect already in progress for {ble_address}."
-                                )
-                            if ble_address in _ble_executor_degraded_addresses:
-                                logger.error(
-                                    "BLE executor degraded for %s: too many orphaned workers. "
-                                    "Reconnect or restart required to restore BLE connectivity.",
-                                    ble_address,
-                                )
-                                raise BleExecutorDegradedError(
-                                    f"BLE executor degraded for {ble_address}; reset required"
-                                )
-                            try:
-                                connect_future = _get_ble_executor().submit(
-                                    connect_iface, iface
-                                )
-                            except RuntimeError as exc:
-                                logger.exception(
-                                    "BLE connect() submission failed for %s",
-                                    ble_address,
-                                )
-                                raise TimeoutError(
-                                    f"BLE connect could not be scheduled for {ble_address}."
-                                ) from exc
-                            _ble_future = connect_future
-                            _ble_future_address = ble_address
-                            _ble_future_started_at = time.monotonic()
-                            _ble_future_timeout_secs = BLE_CONNECT_TIMEOUT_SECS
-                        connect_future.add_done_callback(_clear_ble_future)
-                        try:
-                            _wait_for_future_result_with_shutdown(
-                                connect_future,
-                                timeout_seconds=BLE_CONNECT_TIMEOUT_SECS,
-                            )
-                            logger.info(f"BLE connection established to {ble_address}")
-                            reset_executor_degraded_state(ble_address=ble_address)
-                        except (TimeoutError, FuturesTimeoutError) as err:
-                            if shutting_down or str(err) == "Shutdown in progress":
-                                logger.debug(
-                                    "BLE connect() interrupted by shutdown for %s",
-                                    ble_address,
-                                )
-                                shutdown_iface = iface
-                                if connect_future.cancel():
-                                    _clear_ble_future(connect_future)
-                                else:
-                                    _schedule_ble_future_cleanup(
-                                        connect_future,
-                                        ble_address,
-                                        reason="connect shutdown cancellation",
-                                    )
-                                    _attach_late_ble_interface_disposer(
-                                        connect_future,
-                                        ble_address,
-                                        reason="connect shutdown cancellation",
-                                        fallback_iface=shutdown_iface,
-                                    )
-                                iface = None
-                                meshtastic_iface = None
-                            else:
-                                # Use logger.exception so timeouts include stack context (TRY400),
-                                # but raise a short error and keep operator guidance in logs (TRY003).
-                                ble_connect_timeout_logged_for_attempt = True
-                                logger.exception(
-                                    f"BLE connect() call timed out after {BLE_CONNECT_TIMEOUT_SECS} seconds for %s.",
-                                    ble_address,
-                                )
-                                logger.warning(
-                                    "This may indicate a BlueZ or adapter issue."
-                                )
-                                logger.warning(
-                                    f"BlueZ may be in a bad state. {BLE_TROUBLESHOOTING_GUIDANCE.format(ble_address=ble_address)}"
-                                )
-                                # Best-effort cancellation: a hung BLE connect blocks the worker
-                                # thread, so we cancel to allow retries only if it completes.
-                                if connect_future.cancel():
-                                    _clear_ble_future(connect_future)
-                                else:
-                                    timed_out_iface = iface
-                                    # Clear global/local references before attaching late
-                                    # disposer so late completions cannot observe stale active
-                                    # globals and skip cleanup.
-                                    iface = None
-                                    meshtastic_iface = None
-                                    _schedule_ble_future_cleanup(
-                                        connect_future,
-                                        ble_address,
-                                        reason="connect timeout",
-                                    )
-                                    _attach_late_ble_interface_disposer(
-                                        connect_future,
-                                        ble_address,
-                                        reason="connect timeout",
-                                        fallback_iface=timed_out_iface,
-                                    )
-                                    timeout_count = _record_ble_timeout(ble_address)
-                                    _maybe_reset_ble_executor(
-                                        ble_address, timeout_count
-                                    )
-                                # Don't use iface if connect() timed out - it may be in an inconsistent state
-                                iface = None
-                                meshtastic_iface = None
-                                raise TimeoutError(
-                                    f"BLE connect() timed out for {ble_address}."
-                                ) from err
-                            raise
-                    elif iface is not None and hasattr(iface, "connect"):
-                        logger.debug(
-                            "Skipping explicit BLE connect in compatibility mode; "
-                            "interface is expected to connect during initialization for %s",
-                            ble_address,
-                        )
-
-                    client = iface
-                else:
-                    logger.error("No BLE address provided.")
-                    return None
-
-            elif connection_type == CONNECTION_TYPE_TCP:
-                # TCP connection
-                target_host = config["meshtastic"].get(CONFIG_KEY_HOST)
-                if not target_host:
-                    logger.error(
-                        "No host specified in Meshtastic configuration for TCP connection."
-                    )
-                    return None
-
-                target_port = DEFAULT_TCP_PORT
-                configured_port = config["meshtastic"].get(CONFIG_KEY_PORT)
-                if configured_port is not None:
-                    try:
-                        parsed_port = int(configured_port)
-                        if parsed_port <= 0 or parsed_port > 65535:
-                            raise ValueError
-                        target_port = parsed_port
-                    except (TypeError, ValueError):
-                        logger.warning(
-                            "Invalid meshtastic.port value %r; using default TCP port %s",
-                            configured_port,
-                            DEFAULT_TCP_PORT,
-                        )
-
-                logger.info(f"Connecting to host {target_host}:{target_port}")
-
-                # Connect without progress indicator
-                client = meshtastic.tcp_interface.TCPInterface(
-                    hostname=target_host,
-                    portNumber=target_port,
-                    timeout=configured_timeout_arg,
-                )
-            else:
-                logger.error(f"Unknown connection type: {connection_type}")
-                return None
-
-            successful = True
-
-            # Acquire lock only for the final setup and subscription
-            with meshtastic_lock:
-                # CRITICAL VALIDATION: Verify we're connected to the correct BLE device.
-                # This prevents connection to wrong device due to substring matching
-                # bugs in meshtastic library's find_device() function. The official
-                # version uses substring matching: `address in (device.name, device.address)`
-                # which can match a non-target device if its name contains to
-                # configured address as a substring.
-                #
-                # Example vulnerability scenario:
-                # - Configured address: AA:BB:CC:DD:EE:FF (Meshtastic device)
-                # - Nearby device name: "AA:BB:CC:DD:EE:FF-Sensor" (car/handset/etc)
-                # - Result: Bot incorrectly matches and connects to non-Meshtastic device
-                #
-                # This validation works with both official and forked meshtastic versions.
-                # If validation fails, we disconnect immediately to prevent further issues.
-                if connection_type == CONNECTION_TYPE_BLE:
-                    expected_ble_address = config["meshtastic"].get(
-                        CONFIG_KEY_BLE_ADDRESS
-                    )
-                    if expected_ble_address and not _validate_ble_connection_address(
-                        client, expected_ble_address
-                    ):
-                        # Validation failed - wrong device connected
-                        # Disconnect immediately to prevent communication with wrong device
-                        logger.error(
-                            "BLE connection validation failed - connected to wrong device. "
-                            "Disconnecting and raising error to force retry."
-                        )
-                        was_shared_interface = client is meshtastic_iface
-                        try:
-                            if was_shared_interface:
-                                _disconnect_ble_interface(
-                                    meshtastic_iface, reason="address validation failed"
-                                )
-                            else:
-                                client.close()
-                        except Exception as e:
-                            logger.warning(f"Error closing invalid BLE connection: {e}")
-                        finally:
-                            if was_shared_interface:
-                                meshtastic_iface = None
-                        raise ConnectionRefusedError(
-                            f"Connected to wrong BLE device. Expected: {expected_ble_address}"
-                        )
-
-                # Clear health probe deadlines BEFORE resetting clock skew to prevent
-                # race condition where a late ACK from the previous interface could
-                # seed the new connection with stale skew values.
-                #
-                # Initialize connection timing state exactly once before metadata
-                # probes, node-info fetches, and subscription setup so reconnects
-                # cannot handle inbound packets with stale session timing.
-                with _health_probe_request_lock:
-                    _health_probe_request_deadlines.clear()
-                    with _relay_rx_time_clock_skew_lock:
-                        RELAY_START_TIME = time.time()
-                        _relay_connection_started_monotonic_secs = time.monotonic()
-                        _relay_rx_time_clock_skew_secs = None
-                        if not _startup_packet_drain_applied:
-                            _relay_startup_drain_deadline_monotonic_secs = None
-                            _relay_reconnect_prestart_bootstrap_deadline_monotonic_secs = (
-                                None
-                            )
-                            startup_drain_pending_for_this_connect = True
-                            timing_mode = "startup_pending"
-                        else:
-                            _relay_startup_drain_deadline_monotonic_secs = None
-                            _relay_reconnect_prestart_bootstrap_deadline_monotonic_secs = (
-                                _relay_connection_started_monotonic_secs
-                                + RECONNECT_PRESTART_BOOTSTRAP_WINDOW_SECS
-                            )
-                            reconnect_bootstrap_armed_for_this_connect = True
-                            timing_mode = "reconnect"
-                        logger.debug(
-                            "Initialized connection timing state mode=%s start=%.3f monotonic_start=%.3f startup_drain_deadline=%s reconnect_bootstrap_deadline=%s",
-                            timing_mode,
-                            RELAY_START_TIME,
-                            _relay_connection_started_monotonic_secs,
-                            _relay_startup_drain_deadline_monotonic_secs,
-                            _relay_reconnect_prestart_bootstrap_deadline_monotonic_secs,
-                        )
-
-                # Publish the active client only after per-connection timing state
-                # has been reset, so callbacks cannot observe stale skew windows.
-                if shutting_down:
-                    logger.debug(
-                        "Shutdown started during connect setup; closing new client before publish"
-                    )
-                    try:
-                        if client is meshtastic_iface:
-                            _disconnect_ble_interface(
-                                meshtastic_iface,
-                                reason="connect setup cancelled by shutdown",
-                            )
-                            meshtastic_iface = None
-                        else:
-                            client.close()
-                    except Exception as cleanup_error:  # noqa: BLE001 - best effort
-                        logger.warning(
-                            "Error closing Meshtastic client during shutdown race: %s",
-                            cleanup_error,
-                        )
-                    with _relay_rx_time_clock_skew_lock:
-                        if reconnect_bootstrap_armed_for_this_connect:
-                            _relay_reconnect_prestart_bootstrap_deadline_monotonic_secs = (
-                                None
-                            )
-                    successful = False
-                    return None
-
-                meshtastic_client = client
-                _relay_active_client_id = id(client)
-                client_assigned_for_this_connect = True
-
-                node_info = client.getMyNodeInfo()
-
-                if shutting_down:
-                    logger.debug(
-                        "Shutdown started during connect setup (after getMyNodeInfo); rolling back client"
-                    )
-                    client_assigned_for_this_connect = _rollback_connect_attempt_state(
-                        client=client,
-                        client_assigned_for_this_connect=client_assigned_for_this_connect,
-                        startup_drain_armed_for_this_connect=startup_drain_armed_for_this_connect,
-                        startup_drain_applied_for_this_connect=startup_drain_applied_for_this_connect,
-                        reconnect_bootstrap_armed_for_this_connect=reconnect_bootstrap_armed_for_this_connect,
-                        lock_held=True,
-                    )
-                    successful = False
-                    return None
-
-                # Safely access node info fields
-                user_info = node_info.get("user", {}) if node_info else {}
-                short_name = user_info.get("shortName", "unknown")
-                hw_model = user_info.get("hwModel", "unknown")
-
-                # Get firmware version from device metadata
-                metadata = _get_device_metadata(client)
-                firmware_version = metadata["firmware_version"]
-
-                if shutting_down:
-                    logger.debug(
-                        "Shutdown started during connect setup (after metadata); rolling back client"
-                    )
-                    client_assigned_for_this_connect = _rollback_connect_attempt_state(
-                        client=client,
-                        client_assigned_for_this_connect=client_assigned_for_this_connect,
-                        startup_drain_armed_for_this_connect=startup_drain_armed_for_this_connect,
-                        startup_drain_applied_for_this_connect=startup_drain_applied_for_this_connect,
-                        reconnect_bootstrap_armed_for_this_connect=reconnect_bootstrap_armed_for_this_connect,
-                        lock_held=True,
-                    )
-                    successful = False
-                    return None
-
-                if metadata.get("success"):
-                    logger.info(
-                        f"Connected to {short_name} / {hw_model} / Meshtastic Firmware version {firmware_version}"
-                    )
-                else:
-                    logger.info(f"Connected to {short_name} / {hw_model}")
-                    logger.debug(
-                        "Device firmware version unavailable from getMetadata()"
-                    )
-
-                # Arm startup drain only once setup completes, so the full drain
-                # interval is available when receive handling becomes active.
-                if startup_drain_pending_for_this_connect:
-                    with _relay_rx_time_clock_skew_lock:
-                        if not _startup_packet_drain_applied:
-                            _relay_startup_drain_deadline_monotonic_secs = (
-                                time.monotonic() + STARTUP_PACKET_DRAIN_SECS
-                            )
-                            _startup_packet_drain_applied = True
-                            startup_drain_applied_for_this_connect = True
-                            startup_drain_armed_for_this_connect = True
-                    if startup_drain_armed_for_this_connect:
-                        logger.debug(
-                            "Armed startup drain window deadline=%s after setup completion",
-                            _relay_startup_drain_deadline_monotonic_secs,
-                        )
-
-                if shutting_down:
-                    logger.debug(
-                        "Shutdown started before callback subscription; rolling back client"
-                    )
-                    client_assigned_for_this_connect = _rollback_connect_attempt_state(
-                        client=client,
-                        client_assigned_for_this_connect=client_assigned_for_this_connect,
-                        startup_drain_armed_for_this_connect=startup_drain_armed_for_this_connect,
-                        startup_drain_applied_for_this_connect=startup_drain_applied_for_this_connect,
-                        reconnect_bootstrap_armed_for_this_connect=reconnect_bootstrap_armed_for_this_connect,
-                        lock_held=True,
-                    )
-                    successful = False
-                    return None
-
-                # Subscribe to message and connection-lost events.
-                ensure_meshtastic_callbacks_subscribed()
-
-                _schedule_connect_time_calibration_probe(
-                    client,
-                    connection_type=connection_type,
-                    active_config=config,
-                )
-
-        except ConnectionRefusedError:
-            _rollback_connect_attempt_state(
-                client=client,
-                client_assigned_for_this_connect=client_assigned_for_this_connect,
-                startup_drain_armed_for_this_connect=startup_drain_armed_for_this_connect,
-                startup_drain_applied_for_this_connect=startup_drain_applied_for_this_connect,
-                reconnect_bootstrap_armed_for_this_connect=reconnect_bootstrap_armed_for_this_connect,
-            )
-            client_assigned_for_this_connect = False
-            successful = False
-            return None
-        except (MemoryError, BleExecutorDegradedError):
-            _rollback_connect_attempt_state(
-                client=client,
-                client_assigned_for_this_connect=client_assigned_for_this_connect,
-                startup_drain_armed_for_this_connect=startup_drain_armed_for_this_connect,
-                startup_drain_applied_for_this_connect=startup_drain_applied_for_this_connect,
-                reconnect_bootstrap_armed_for_this_connect=reconnect_bootstrap_armed_for_this_connect,
-            )
-            client_assigned_for_this_connect = False
-            successful = False
-            logger.exception("Critical connection error")
-            return None
-        except (FuturesTimeoutError, TimeoutError) as e:
-            successful = False
-            client_assigned_for_this_connect = _rollback_connect_attempt_state(
-                client=client,
-                client_assigned_for_this_connect=client_assigned_for_this_connect,
-                startup_drain_armed_for_this_connect=startup_drain_armed_for_this_connect,
-                startup_drain_applied_for_this_connect=startup_drain_applied_for_this_connect,
-                reconnect_bootstrap_armed_for_this_connect=reconnect_bootstrap_armed_for_this_connect,
-            )
-            if shutting_down:
-                break
-            if (
-                connection_type == CONNECTION_TYPE_BLE
-                and ble_address
-                and not ble_connect_timeout_logged_for_attempt
-                and str(e).startswith("BLE connect() timed out for ")
-            ):
-                logger.exception(
-                    f"BLE connect() call timed out after {BLE_CONNECT_TIMEOUT_SECS} seconds for %s.",
-                    ble_address,
-                )
-                logger.warning("This may indicate a BlueZ or adapter issue.")
-                logger.warning(
-                    f"BlueZ may be in a bad state. {BLE_TROUBLESHOOTING_GUIDANCE.format(ble_address=ble_address)}"
-                )
-            attempts += 1
-            if retry_limit == INFINITE_RETRIES:
-                timeout_attempts += 1
-                if timeout_attempts > MAX_TIMEOUT_RETRIES_INFINITE:
-                    logger.exception(
-                        "Connection timed out after %s attempts (unlimited retries); aborting",
-                        attempts,
-                    )
-                    return None
-            elif attempts > retry_limit:
-                logger.exception("Connection failed after %s attempts", attempts)
-                return None
-
-            wait_time = _get_connection_retry_wait_time(attempts)
-            logger.warning(
-                "Connection attempt %s timed out (%s). Retrying in %s seconds...",
-                attempts,
-                e,
-                wait_time,
-            )
-            time.sleep(wait_time)
-        except Exception as e:
-            successful = False
-            client_assigned_for_this_connect = _rollback_connect_attempt_state(
-                client=client,
-                client_assigned_for_this_connect=client_assigned_for_this_connect,
-                startup_drain_armed_for_this_connect=startup_drain_armed_for_this_connect,
-                startup_drain_applied_for_this_connect=startup_drain_applied_for_this_connect,
-                reconnect_bootstrap_armed_for_this_connect=reconnect_bootstrap_armed_for_this_connect,
-            )
-            if shutting_down:
-                logger.debug("Shutdown in progress. Aborting connection attempts.")
-                break
-            if (
-                connection_type == CONNECTION_TYPE_BLE
-                and ble_address
-                and _is_ble_duplicate_connect_suppressed_error(e)
-            ):
-                logger.warning(
-                    "Detected duplicate BLE connect suppression for %s",
-                    ble_address,
-                )
-                if not _reset_ble_connection_gate_state(
-                    ble_address,
-                    reason="duplicate connect suppression",
-                ):
-                    logger.debug(
-                        "BLE gate reset hook unavailable for %s; retrying without local reset",
-                        ble_address,
-                    )
-            attempts += 1
-            if (
-                connection_type == CONNECTION_TYPE_BLE
-                and ble_address
-                # Keep discovery-triggered scan recovery scoped to compatibility
-                # mode; forked auto_reconnect-capable implementations handle
-                # discovery/connect retries internally.
-                and not supports_auto_reconnect
-                and _is_ble_discovery_error(e)
-            ):
-                ble_scan_after_failure = True
-                ble_scan_reason = type(e).__name__
-            if retry_limit == 0 or attempts <= retry_limit:
-                wait_time = _get_connection_retry_wait_time(attempts)
-                logger.warning(
-                    "An unexpected error occurred on attempt %s: %s. Retrying in %s seconds...",
-                    attempts,
-                    e,
-                    wait_time,
-                )
-                time.sleep(wait_time)
-            else:
-                logger.exception("Connection failed after %s attempts", attempts)
-                return None
-
-    return meshtastic_client
-
-
-def on_lost_meshtastic_connection(
-    interface: Any = None,
-    detection_source: str = "unknown",
-    topic: Any = pub.AUTO_TOPIC,
-) -> None:
-    """
-    Mark the Meshtastic connection as lost, close the current client, and start an asynchronous reconnect.
-
-    If a shutdown is underway or a reconnect is already in progress this function returns immediately. When proceeding it sets the module-level `reconnecting` flag, attempts a best-effort close/cleanup of the current Meshtastic client/interface (with special handling for BLE interfaces), clears any in-flight BLE future state, and schedules the `reconnect()` coroutine on the global event loop.
-
-    Parameters:
-        detection_source (str): Identifier for where or how the loss was detected; if `"unknown"`, the function will prefer an interface-provided `_last_disconnect_source`, then derive a name from `topic`, and finally fall back to `"meshtastic.connection.lost"`.
-        topic (Any): Optional pubsub topic object (from pypubsub); when provided and `detection_source` is `"unknown"`, the topic's name will be used to derive the detection source.
-    """
-    # Keep these as one-global-per-line to minimize merge churn as this list evolves.
-    global meshtastic_client
-    global _relay_active_client_id
-    global meshtastic_iface
-    global reconnecting
-    global shutting_down
-    global event_loop
-    global reconnect_task
-    global _ble_future
-    global _ble_future_address
-    global _ble_future_started_at
-    global _ble_future_timeout_secs
-    global _ble_executor
-
-    with meshtastic_lock:
-        if shutting_down:
-            logger.debug("Shutdown in progress. Not attempting to reconnect.")
-            return
-        active_client = meshtastic_client
-        active_client_id = _relay_active_client_id
-        if (
-            interface is not None
-            and active_client is None
-            and (
-                _callbacks_tearing_down
-                or subscribed_to_connection_lost
-                or shutting_down
-            )
-        ):
-            logger.debug(
-                "Ignoring connection-lost event because no Meshtastic interface is currently active"
-            )
-            return
-        if interface is not None and active_client is not None:
-            expected_client_id = (
-                active_client_id if active_client_id is not None else id(active_client)
-            )
-            if id(interface) != expected_client_id:
-                logger.debug(
-                    "Ignoring connection-lost event from stale Meshtastic interface "
-                    "(event_interface_id=%s active_client_id=%s)",
-                    id(interface),
-                    expected_client_id,
-                )
-                return
-        if reconnecting:
-            logger.debug(
-                "Reconnection already in progress. Skipping additional reconnection attempt."
-            )
-            return
-        if detection_source == "unknown":
-            interface_source = getattr(interface, "_last_disconnect_source", None)
-            if isinstance(interface_source, str) and (
-                stripped := interface_source.strip()
-            ):
-                # Strip 'ble.' prefix to make detection source library-agnostic
-                res = stripped[4:].strip() if stripped.startswith("ble.") else stripped
-                if res:
-                    detection_source = res
-                    logger.debug(
-                        "Using interface-provided detection source: %s",
-                        detection_source,
-                    )
-
-            if detection_source == "unknown":
-                if topic is not None and topic is not pub.AUTO_TOPIC:
-                    # Real topic object from pypubsub - extract its name
-                    detection_source = getattr(topic, "getName", lambda: str(topic))()
-                    logger.debug(
-                        "Using pubsub topic-derived detection source: %s",
-                        detection_source,
-                    )
-                else:
-                    # Called directly without a topic, or with AUTO_TOPIC sentinel
-                    logger.debug(
-                        "_last_disconnect_source unavailable; using default detection source"
-                    )
-                    detection_source = "meshtastic.connection.lost"
-
-        reconnecting = True
-        logger.error(f"Lost connection ({detection_source}). Reconnecting...")
-
-        if meshtastic_client:
-            if meshtastic_client is meshtastic_iface:
-                # This is a BLE interface - use proper disconnect sequence
-                logger.debug("Disconnecting BLE interface due to connection loss")
-                _disconnect_ble_interface(
-                    meshtastic_iface, reason=f"connection loss: {detection_source}"
-                )
-                meshtastic_iface = None
-            else:
-                # Serial or TCP interface - use standard close()
-                try:
-                    meshtastic_client.close()
-                except OSError as e:
-                    if e.errno == ERRNO_BAD_FILE_DESCRIPTOR:
-                        # Bad file descriptor, already closed
-                        pass
-                    else:
-                        logger.warning(f"Error closing Meshtastic client: {e}")
-                except Exception as e:
-                    logger.warning(f"Error closing Meshtastic client: {e}")
-        meshtastic_client = None
-        _relay_active_client_id = None
-        ble_future_to_cancel = None
-        stale_executor = None
-        stale_ble_address: str | None = None
-        with _ble_executor_lock:
-            stale_ble_address = _ble_future_address
-            if _ble_future and not _ble_future.done():
-                logger.debug(
-                    "Clearing stale BLE future before reconnect (%s)",
-                    detection_source,
-                )
-                ble_future_to_cancel = _ble_future
-                _ble_future = None
-                if _ble_future_address:
-                    with _ble_timeout_lock:
-                        _ble_timeout_counts.pop(_ble_future_address, None)
-                _ble_future_address = None
-                _ble_future_started_at = None
-                _ble_future_timeout_secs = None
-                if _ble_executor is not None:
-                    stale_executor = _ble_executor
-                    _ble_executor = ThreadPoolExecutor(max_workers=1)
-
-        if ble_future_to_cancel is not None:
-            if stale_ble_address:
-                _attach_late_ble_interface_disposer(
-                    ble_future_to_cancel,
-                    stale_ble_address,
-                    reason=f"connection loss: {detection_source}",
-                )
-            ble_future_to_cancel.cancel()
-        if stale_executor is not None:
-            try:
-                stale_executor.shutdown(wait=False, cancel_futures=True)
-            except TypeError:
-                stale_executor.shutdown(wait=False)
-
-        if stale_ble_address is not None:
-            reset_executor_degraded_state(ble_address=stale_ble_address)
-        else:
-            should_reset_all_degraded = False
-            with _ble_executor_lock:
-                if _ble_executor_degraded_addresses:
-                    should_reset_all_degraded = True
-                    logger.debug(
-                        "Resetting degraded BLE executor state during reconnect "
-                        "(no stale_ble_address but degraded addresses exist)"
-                    )
-            if should_reset_all_degraded:
-                reset_executor_degraded_state(reset_all=True)
-
-        if event_loop and not event_loop.is_closed():
-            reconnect_task = event_loop.create_task(reconnect())
-
-
-async def reconnect() -> None:
-    """
-    Re-establish the Meshtastic connection using exponential backoff.
-
-    Retries connect_meshtastic(force_connect=True) until a connection is obtained, the application begins shutting down, or the task is cancelled. Starts with DEFAULT_BACKOFF_TIME and doubles the wait after each failed attempt, capped at 300 seconds. Stops promptly on cancellation or when shutting_down is set, and ensures the module-level `reconnecting` flag is cleared before returning.
-    """
-    global reconnecting, shutting_down, reconnect_task_future
-    backoff_time = DEFAULT_BACKOFF_TIME
-    try:
-        while not shutting_down:
-            try:
-                logger.info(
-                    f"Reconnection attempt starting in {backoff_time} seconds..."
-                )
-
-                # Show reconnection countdown with Rich (if not in a service)
-                if not is_running_as_service():
-                    try:
-                        from rich.progress import (
-                            BarColumn,
-                            Progress,
-                            TextColumn,
-                            TimeRemainingColumn,
-                        )
-                    except ImportError:
-                        logger.debug(
-                            "Rich not available; falling back to simple reconnection delay"
-                        )
-                        await asyncio.sleep(backoff_time)
-                    else:
-                        with Progress(
-                            TextColumn("[cyan]Meshtastic: Reconnecting in"),
-                            BarColumn(),
-                            TextColumn("[cyan]{task.percentage:.0f}%"),
-                            TimeRemainingColumn(),
-                            transient=True,
-                        ) as progress:
-                            task = progress.add_task("Waiting", total=backoff_time)
-                            for _ in range(backoff_time):
-                                if shutting_down:
-                                    break
-                                await asyncio.sleep(1)
-                                progress.update(task, advance=1)
-                else:
-                    await asyncio.sleep(backoff_time)
-                if shutting_down:
-                    logger.debug(
-                        "Shutdown in progress. Aborting reconnection attempts."
-                    )
-                    break
-                loop = asyncio.get_running_loop()
-                # Pass the current config during reconnection to ensure matrix_rooms is populated
-                # Using None for passed_config would skip matrix_rooms initialization
-                connect_future = asyncio.ensure_future(
-                    loop.run_in_executor(None, connect_meshtastic, config, True)
-                )
-                reconnect_task_future = connect_future
-                connected_client = await connect_future
-                if connected_client is not None:
-                    logger.info("Reconnected successfully.")
-                    break
-            except Exception:
-                if shutting_down:
-                    break
-                logger.exception("Reconnection attempt failed")
-                backoff_time = min(backoff_time * 2, 300)  # Cap backoff at 5 minutes
-    except asyncio.CancelledError:
-        logger.info("Reconnection task was cancelled.")
-    finally:
-        if reconnect_task_future is not None:
-            if reconnect_task_future.done() and not reconnect_task_future.cancelled():
-                reconnect_task_future = None
-        reconnecting = False
-
-
-def on_meshtastic_message(packet: dict[str, Any], interface: Any) -> None:
-    """
-    Route an incoming Meshtastic packet to configured Matrix rooms or installed plugins based on runtime configuration.
-
-    Processes the decoded packet and, depending on interaction settings and packet contents, will relay emoji reactions and replies to mapped Matrix events, dispatch ordinary text messages to Matrix rooms mapped to the packet's channel (unless the message is a direct message to the relay node or handled by a plugin), and hand non-text or unhandled packets to installed plugins with a per-plugin timeout.
-
-    Parameters:
-        packet (dict): Decoded Meshtastic packet. Expected keys include:
-            - 'decoded' (dict): may contain 'text', 'replyId', 'portnum', and optional 'emoji'
-            - 'fromId' or 'from' (sender id)
-            - 'to' (destination id)
-            - 'id' (packet id)
-            - optional 'channel' (mapped channel value)
-        interface: Meshtastic interface used to resolve node information and the relay node id. Must provide .myInfo.my_node_num and a .nodes mapping for sender metadata.
-    """
-    global _relay_rx_time_clock_skew_secs, _relay_startup_drain_deadline_monotonic_secs
-
-    # Validate packet structure
-    if not packet or not isinstance(packet, dict):
-        logger.error("Received malformed packet: packet is None or not a dict")
-        return
-
-    if shutting_down:
-        logger.debug("Shutdown in progress. Ignoring incoming messages.")
-        return
-
-    # Read-mostly guard values; avoid meshtastic_lock here because callbacks can
-    # fire synchronously while connect_meshtastic() still holds that lock.
-    active_client = meshtastic_client
-    active_client_id = _relay_active_client_id
-    if active_client is None:
-        if active_client_id is not None:
-            logger.error(
-                "Inconsistent relay state: active_client is None but active_client_id=%s — this should not happen",
-                active_client_id,
-            )
-        # Runtime callbacks can still arrive briefly during reconnect/teardown
-        # windows because subscriptions are process-lifetime.
-        #
-        # Keep direct unit-level handler invocation behavior unchanged when no
-        # active session is being transitioned.
-        if (
-            _callbacks_tearing_down
-            or subscribed_to_messages
-            or reconnecting
-            or shutting_down
-        ):
-            logger.debug(
-                "Ignoring packet because no Meshtastic interface is currently active"
-            )
-            return
-    else:
-        expected_client_id = (
-            active_client_id if active_client_id is not None else id(active_client)
-        )
-        if id(interface) != expected_client_id:
-            logger.debug(
-                "Ignoring packet from stale Meshtastic interface (packet_interface_id=%s active_client_id=%s)",
-                id(interface),
-                expected_client_id,
-            )
-            return
-
-    # Parse rxTime early so health-probe responses can calibrate packet clock skew.
-    rx_time_raw = packet.get("rxTime", 0)
-    try:
-        rx_time = float(rx_time_raw)
-    except (TypeError, ValueError):
-        rx_time = 0
-
-    is_health_probe_response = _claim_health_probe_response_and_maybe_calibrate(
-        packet, interface, rx_time
-    )
-    if is_health_probe_response:
-        decoded = packet.get("decoded")
-        portnum = decoded.get("portnum") if isinstance(decoded, dict) else None
-        logger.debug(
-            "[HEALTH_CHECK] Metadata probe response requestId=%s from=%s port=%s",
-            _extract_packet_request_id(packet),
-            packet.get("fromId") or packet.get("from"),
-            _get_portnum_name(portnum),
-        )
-        return
-
-    now_monotonic = time.monotonic()
-    with _relay_rx_time_clock_skew_lock:
-        relay_start_time = RELAY_START_TIME
-        startup_drain_deadline = _relay_startup_drain_deadline_monotonic_secs
-
-    if startup_drain_deadline is not None:
-        remaining_drain_secs = startup_drain_deadline - now_monotonic
-        if remaining_drain_secs > 0:
-            calibrated_during_drain = False
-            if rx_time > 0:
-                calibrated_during_drain = _seed_connect_time_skew(rx_time)
-            if calibrated_during_drain and rx_time < relay_start_time:
-                logger.debug(
-                    "Consumed startup bootstrap packet with rxTime %s to calibrate clock skew",
-                    rx_time,
-                )
-            logger.debug(
-                "Dropping inbound packet during startup drain window (remaining=%.3f seconds)",
-                remaining_drain_secs,
-            )
-            return
-        with _relay_rx_time_clock_skew_lock:
-            if (
-                _relay_startup_drain_deadline_monotonic_secs is not None
-                and _relay_startup_drain_deadline_monotonic_secs <= now_monotonic
-            ):
-                _relay_startup_drain_deadline_monotonic_secs = None
-                logger.debug("Startup drain window has ended — accepting packets")
-
-    # Seed clock skew from the first non-health-probe packet with a valid
-    # rxTime so that the cutoff works even when health checks are disabled.
-    # During startup we allow one bounded bootstrap from a pre-start packet,
-    # then consume that packet so backlog is not relayed.
-    calibrated_from_packet = False
-    if rx_time > 0:
-        calibrated_from_packet = _seed_connect_time_skew(rx_time)
-
-    if calibrated_from_packet and rx_time < relay_start_time:
-        logger.debug(
-            "Consumed startup bootstrap packet with rxTime %s to calibrate clock skew",
-            rx_time,
-        )
-        return
-
-    # Filter out old messages (from before relay start) to prevent flooding.
-    # This handles cases where the node dumps stored history upon connection.
-    # When health probes calibrate packet clock skew, adjust the relay start
-    # cutoff so clock offsets do not hide fresh traffic.
-    with _relay_rx_time_clock_skew_lock:
-        relay_start_time = RELAY_START_TIME
-        calibrated_skew = _relay_rx_time_clock_skew_secs
-    effective_relay_start_time = relay_start_time
-    if calibrated_skew is not None:
-        effective_relay_start_time = relay_start_time - calibrated_skew
-
-    if rx_time > 0 and rx_time < effective_relay_start_time:
-        if calibrated_skew is None:
-            logger.debug(
-                "Ignoring old packet with rxTime %s (older than start time %s)",
-                rx_time,
-                relay_start_time,
-            )
-        else:
-            logger.debug(
-                "Ignoring old packet with rxTime %s (older than adjusted start time %s; raw start=%s skew=%s)",
-                rx_time,
-                effective_relay_start_time,
-                relay_start_time,
-                calibrated_skew,
-            )
-        return
-
-    # Full packet logging for debugging (when enabled in config)
-    # Check if full packet logging is enabled - accepts boolean True or string "true"
-    debug_settings = config.get("logging", {}).get("debug", {}) if config else {}
-    full_packets_setting = debug_settings.get("full_packets")
-    if full_packets_setting is True or (
-        isinstance(full_packets_setting, str) and full_packets_setting.lower() == "true"
-    ):
-        logger.debug("Full packet: %s", packet)
-
-    # Log that we received a message (without the full packet details)
-    decoded = packet.get("decoded")
-    if decoded and isinstance(decoded, dict) and decoded.get("text"):
-        logger.info(f"Received Meshtastic message: {decoded.get('text')}")
-    else:
-        portnum = (
-            decoded.get("portnum") if decoded and isinstance(decoded, dict) else None
-        )
-        portnum_name = _get_portnum_name(portnum)
-        from_id = packet.get("fromId") or packet.get("from")
-        from_display = ""
-        if from_id is not None:
-            from_display = _get_node_display_name(from_id, interface, fallback="")
-        details_map = {
-            "from": from_id,
-            "channel": packet.get("channel"),
-            "id": packet.get("id"),
-        }
-        details_map.update(_get_packet_details(decoded, packet, portnum_name))
-
-        details = []
-        if from_display:
-            details.append(from_display)
-        for key, value in details_map.items():
-            if value is not None:
-                if key == "from":
-                    details.append(f"from={value}")
-                elif key == "batt":
-                    details.append(f"{value}")
-                elif key == "voltage":
-                    details.append(f"v={value}")
-                elif key == "temp":
-                    details.append(f"t={value}")
-                elif key == "humidity":
-                    details.append(f"h={value}")
-                elif key == "signal":
-                    details.append(f"s={value}")
-                elif key == "relayed":
-                    details.append(f"r={value}")
-                elif key == "priority":
-                    details.append(f"p={value}")
-                else:
-                    details.append(f"{key}={value}")
-
-        prefix = f"[{portnum_name}] " + " ".join(details)
-        logger.debug(prefix)
-
-    # Check if config is available
-    if config is None:
-        logger.error("No configuration available. Cannot process Meshtastic message.")
-        return
-
-    # Import the configuration helpers
-    from mmrelay.matrix_utils import get_interaction_settings
-
-    # Get interaction settings
-    interactions = get_interaction_settings(config)
-
-    # Filter packets based on interaction settings
-    if packet.get("decoded", {}).get("portnum") == TEXT_MESSAGE_APP:
-        decoded = packet.get("decoded", {})
-        # Filter out reactions if reactions are disabled
-        if (
-            not interactions["reactions"]
-            and decoded.get("replyId") is not None
-            and "emoji" in decoded
-            and decoded.get("emoji") == EMOJI_FLAG_VALUE
-        ):
-            logger.debug(
-                "Filtered out reaction packet due to reactions being disabled."
-            )
-            return
-
-    from mmrelay.matrix_utils import matrix_relay
-
-    global event_loop
-
-    if shutting_down:
-        logger.debug("Shutdown in progress. Ignoring incoming messages.")
-        return
-
-    if event_loop is None:
-        logger.error("Event loop is not set. Cannot process message.")
-        return
-
-    loop = event_loop
-
-    sender = packet.get("fromId") or packet.get("from")
-    toId = packet.get("to")
-
-    decoded = packet.get("decoded", {})
-    text = decoded.get("text")
-    replyId = decoded.get("replyId")
-    emoji_flag = "emoji" in decoded and decoded["emoji"] == EMOJI_FLAG_VALUE
-
-    # Determine if this is a direct message to the relay node
-    from meshtastic.mesh_interface import BROADCAST_NUM
-
-    if not getattr(interface, "myInfo", None):
-        logger.warning("Meshtastic interface missing myInfo; cannot determine node id")
-        return
-    myId = interface.myInfo.my_node_num
-
-    if toId == myId:
-        is_direct_message = True
-    elif toId == BROADCAST_NUM or toId is None:
-        is_direct_message = False
-    else:
-        logger.debug(
-            "Ignoring message intended for node %s (not broadcast or relay).", toId
-        )
-        return
-
-    meshnet_name = config[CONFIG_SECTION_MESHTASTIC][CONFIG_KEY_MESHNET_NAME]
-
-    # Reaction handling (Meshtastic -> Matrix)
-    # If replyId and emoji_flag are present and reactions are enabled, we relay as text reactions in Matrix
-    if replyId and emoji_flag and interactions["reactions"]:
-        longname = _get_name_safely(get_longname, sender)
-        shortname = _get_name_safely(get_shortname, sender)
-        orig = get_message_map_by_meshtastic_id(replyId)
-        if orig:
-            # orig = (matrix_event_id, matrix_room_id, meshtastic_text, meshtastic_meshnet)
-            matrix_event_id, matrix_room_id, meshtastic_text, meshtastic_meshnet = orig
-            abbreviated_text = (
-                meshtastic_text[:40] + "..."
-                if len(meshtastic_text) > 40
-                else meshtastic_text
-            )
-
-            # Import the matrix prefix function
-            from mmrelay.matrix_utils import get_matrix_prefix
-
-            # Get the formatted prefix for the reaction
-            prefix = get_matrix_prefix(config, longname, shortname, meshnet_name)
-
-            reaction_symbol = text.strip() if (text and text.strip()) else "⚠️"
-            reaction_message = (
-                f'\n {prefix}reacted {reaction_symbol} to "{abbreviated_text}"'
-            )
-
-            # Relay the reaction as emote to Matrix, preserving the original meshnet name
-            _fire_and_forget(
-                matrix_relay(
-                    matrix_room_id,
-                    reaction_message,
-                    longname,
-                    shortname,
-                    meshnet_name,
-                    decoded.get("portnum"),
-                    meshtastic_id=packet.get("id"),
-                    meshtastic_replyId=replyId,
-                    meshtastic_text=meshtastic_text,
-                    emote=True,
-                    emoji=True,
-                ),
-                loop=loop,
-            )
-            return
-        else:
-            # Original message not found - fall through to normal text handling
-            # This can happen with:
-            # - Replies to messages from before the relay started
-            # - Cross-meshnet replies where original not in our DB
-            # - Signed IDs that don't match (packet from another node/source)
-            logger.warning(
-                "Original message for reaction (replyId=%s) not found in DB. "
-                "Relaying as normal message instead.",
-                replyId,
-            )
-
-    # Reply handling (Meshtastic -> Matrix)
-    # If replyId is present but emoji is not (or not 1), this is a reply
-    if replyId and not emoji_flag and interactions["replies"]:
-        longname = _get_name_safely(get_longname, sender)
-        shortname = _get_name_safely(get_shortname, sender)
-        orig = get_message_map_by_meshtastic_id(replyId)
-        if orig:
-            # orig = (matrix_event_id, matrix_room_id, meshtastic_text, meshtastic_meshnet)
-            matrix_event_id, matrix_room_id, meshtastic_text, meshtastic_meshnet = orig
-
-            # Import the matrix prefix function
-            from mmrelay.matrix_utils import get_matrix_prefix
-
-            # Get the formatted prefix for the reply
-            prefix = get_matrix_prefix(config, longname, shortname, meshnet_name)
-            formatted_message = f"{prefix}{text}"
-
-            logger.info(f"Relaying Meshtastic reply from {longname} to Matrix")
-
-            # Relay the reply to Matrix with proper reply formatting
-            _fire_and_forget(
-                matrix_relay(
-                    matrix_room_id,
-                    formatted_message,
-                    longname,
-                    shortname,
-                    meshnet_name,
-                    decoded.get("portnum"),
-                    meshtastic_id=packet.get("id"),
-                    meshtastic_replyId=replyId,
-                    meshtastic_text=text,
-                    reply_to_event_id=matrix_event_id,
-                ),
-                loop=loop,
-            )
-            return
-        else:
-            # Original message not found - fall through to normal text handling
-            # This can happen with:
-            # - Replies to messages from before the relay started
-            # - Cross-meshnet replies where original not in our DB
-            # - Signed IDs that don't match (packet from another node/source)
-            logger.warning(
-                "Original message for reply (replyId=%s) not found in DB. "
-                "Relaying as normal message instead.",
-                replyId,
-            )
-
-    # Normal text messages or detection sensor messages
-    if text:
-        # Determine the channel for this message
-        channel = packet.get("channel")
-        if channel is None:
-            # If channel not specified, deduce from portnum
-            # Note: meshtastic-python emits enum names (e.g., "TEXT_MESSAGE_APP") in decoded dicts,
-            # while other paths (protobuf/raw) surface numeric portnums. Support both to avoid drops.
-            if decoded.get("portnum") in (
-                PORTNUM_TEXT_MESSAGE_APP,
-                PORTNUM_DETECTION_SENSOR_APP,
-                TEXT_MESSAGE_APP,
-                DETECTION_SENSOR_APP,
-            ):
-                channel = DEFAULT_CHANNEL_VALUE
-            else:
-                logger.debug(
-                    f"Unknown portnum {decoded.get('portnum')}, cannot determine channel"
-                )
-                return
-
-        # Normalize channel to integer to prevent type mismatch issues
-        try:
-            channel = int(channel)
-        except (ValueError, TypeError):
-            logger.warning(
-                f"Invalid channel value {channel!r} (type: {type(channel).__name__}), "
-                f"defaulting to {DEFAULT_CHANNEL_VALUE}"
-            )
-            channel = DEFAULT_CHANNEL_VALUE
-
-        # Check if channel is mapped to a Matrix room
-        channel_mapped = False
-        iterable_rooms = (
-            matrix_rooms.values() if isinstance(matrix_rooms, dict) else matrix_rooms
-        )
-        for room in iterable_rooms:
-            if not isinstance(room, dict):
-                continue
-
-            room_channel = _normalize_room_channel(room)
-            if room_channel is None:
-                continue
-
-            if room_channel == channel:
-                channel_mapped = True
-                logger.debug(
-                    f"Channel {channel} mapped to Matrix room {room.get('id', 'unknown')}"
-                )
-                break
-
-        if not channel_mapped:
-            # Use WARNING level so this is visible without debug logging enabled
-            # This helps users diagnose configuration issues
-            available_channels = []
-            for room in iterable_rooms:
-                if isinstance(room, dict):
-                    ch = _normalize_room_channel(room)
-                    if ch is not None:
-                        available_channels.append(ch)
-
-            logger.warning(
-                f"Skipping message from unmapped channel {channel}. "
-                f"Available channels in config: {available_channels}. "
-                f"Check your matrix_rooms configuration to ensure this channel is mapped."
-            )
-            return
-
-        # If detection_sensor is disabled and this is a detection sensor packet, skip it
-        portnum = decoded.get("portnum")
-        if (
-            portnum == PORTNUM_DETECTION_SENSOR_APP or portnum == DETECTION_SENSOR_APP
-        ) and not get_meshtastic_config_value(
-            config, "detection_sensor", DEFAULT_DETECTION_SENSOR
-        ):
-            logger.debug(
-                "Detection sensor packet received, but detection sensor processing is disabled."
-            )
-            return
-
-        # Attempt to get longname/shortname from database or nodes
-        longname = _get_name_or_none(get_longname, sender)
-        if longname is None:
-            logger.debug(
-                "Failed to get longname from database for %s, will try interface fallback",
-                sender,
-            )
-
-        shortname = _get_name_or_none(get_shortname, sender)
-        if shortname is None:
-            logger.debug(
-                "Failed to get shortname from database for %s, will try interface fallback",
-                sender,
-            )
-
-        if not longname or not shortname:
-            node = interface.nodes.get(sender)
-            if node:
-                user = node.get("user")
-                if user:
-                    if not longname:
-                        longname_val = user.get("longName")
-                        if longname_val and sender is not None:
-                            save_longname(sender, longname_val)
-                            longname = longname_val
-                    if not shortname:
-                        shortname_val = user.get("shortName")
-                        if shortname_val and sender is not None:
-                            save_shortname(sender, shortname_val)
-                            shortname = shortname_val
-            else:
-                logger.debug(f"Node info for sender {sender} not available yet.")
-
-        # If still not available, fallback to sender ID
-        if not longname:
-            longname = str(sender)
-        if not shortname:
-            shortname = str(sender)
-
-        # Import the matrix prefix function
-        from mmrelay.matrix_utils import get_matrix_prefix
-
-        # Get the formatted prefix
-        prefix = get_matrix_prefix(config, longname, shortname, meshnet_name)
-        formatted_message = f"{prefix}{text}"
-
-        # Plugin functionality - Check if any plugin handles this message before relaying
-        found_matching_plugin = _run_meshtastic_plugins(
-            packet=packet,
-            formatted_message=formatted_message,
-            longname=longname,
-            meshnet_name=meshnet_name,
-            loop=loop,
-            cfg=config,
-        )
-
-        # If message is a DM or handled by plugin, do not relay further
-        if is_direct_message:
-            logger.debug(
-                f"Received a direct message from {longname}: {text}. Not relaying to Matrix."
-            )
-            return
-        if found_matching_plugin:
-            logger.debug("Message was handled by a plugin. Not relaying to Matrix.")
-            return
-
-        # Check if matrix_rooms is empty BEFORE attempting to relay
-        # This can happen during startup race conditions where messages arrive
-        # before matrix_rooms is populated, or during reconnection
-        if not matrix_rooms:
-            logger.warning(
-                f"matrix_rooms is empty - cannot relay message from {longname}. "
-                f"This may indicate a startup race condition or configuration issue. "
-                f"Message will be dropped: {text[:50]}{'...' if len(text) > 50 else ''}"
-            )
-            return
-
-        # Relay the message to all Matrix rooms mapped to this channel
-        logger.info(f"Relaying Meshtastic message from {longname} to Matrix")
-
-        iterable_rooms = (
-            matrix_rooms.values() if isinstance(matrix_rooms, dict) else matrix_rooms
-        )
-        for room in iterable_rooms:
-            if not isinstance(room, dict):
-                continue
-
-            room_channel = _normalize_room_channel(room)
-            if room_channel is None:
-                continue
-
-            if room_channel == channel:
-                # Storing the message_map (if enabled) occurs inside matrix_relay() now,
-                # controlled by relay_reactions.
-                try:
-                    _fire_and_forget(
-                        matrix_relay(
-                            room["id"],
-                            formatted_message,
-                            longname,
-                            shortname,
-                            meshnet_name,
-                            decoded.get("portnum"),
-                            meshtastic_id=packet.get("id"),
-                            meshtastic_text=text,
-                        ),
-                        loop=loop,
-                    )
-                except Exception:
-                    logger.exception("Error relaying message to Matrix")
-    else:
-        # Non-text messages via plugins
-        portnum = decoded.get("portnum")
-        _run_meshtastic_plugins(
-            packet=packet,
-            formatted_message=None,
-            longname=None,
-            meshnet_name=None,
-            loop=loop,
-            cfg=config,
-            use_keyword_args=True,
-            log_with_portnum=True,
-            portnum=portnum,
-        )
+# Connection lifecycle — implemented in meshtastic.connection, re-exported here
+# for backward-compatible patch targets (tests patch mmrelay.meshtastic_utils.*).
+# Event handlers — implemented in meshtastic.events, re-exported here
+# for backward-compatible patch targets (tests patch mmrelay.meshtastic_utils.*).
 
 
 # Health monitoring config interpretation — implemented in meshtastic.health,
